requirements.txt
# Core dependencies
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0

# LLM API
openai==1.3.0  # For OpenRouter compatibility

# Social Media APIs
tweepy==4.14.0
praw==7.7.1

# Database and logging
sqlite3  # Built-in with Python
APScheduler==3.10.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3
